Move buffered transmits to dedicated module
