+++ rust/connlib/tunnel/src/tests.rs
@@ -2,6 +2,7 @@ use crate::tests::sut::TunnelTest;
 use proptest::test_runner::Config;
 
 mod assertions;
+mod buffered_transmits;
 mod composite_strategy;
 mod flux_capacitor;
 mod reference;
